tools: Move infer_cpu to __init__.py
